refactor: remove unused exception imports and add new exception classes for Alumno and Materia
