+++ src/main/java/facultad/MateriaRepetidaException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package facultad;
 
 public class MateriaRepetidaException extends RuntimeException {
