+++ src/main/java/facultad/Profesor.java
@@ -1,7 +1,5 @@
 package facultad;
 
-import exceptions.MateriaRepetidaException;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
